Refactor inet_ntopSS() and inet_ntopAI() into common code, simplify upsclient and upsd
