+++ server/upsd.c
@@ -170,67 +170,6 @@ static int	reload_flag = 0, exit_flag = 0;
 # define SERVICE_UNIT_NAME "nut-server.service"
 #endif
 
-static const char *inet_ntopW(struct sockaddr_storage *s)
-{
-	static char str[40];
-
-	if (!s) {
-		errno = EINVAL;
-		return NULL;
-	}
-
-	switch (s->ss_family)
-	{
-	case AF_INET:
-		return inet_ntop (AF_INET, &(((struct sockaddr_in *)s)->sin_addr), str, 16);
-	case AF_INET6:
-		return inet_ntop (AF_INET6, &(((struct sockaddr_in6 *)s)->sin6_addr), str, 40);
-	default:
-		errno = EAFNOSUPPORT;
-		return NULL;
-	}
-}
-
-static const char *inet_ntopAI(struct addrinfo *ai)
-{
-	/* Note: below we manipulate copies of ai - cannot cast into
-	 * specific structure type pointers right away because:
-	 *   error: cast from 'struct sockaddr *' to 'struct sockaddr_storage *'
-	 *          increases required alignment from 2 to 8
-	 */
-	/* https://stackoverflow.com/a/29147085/4715872
-	 * obviously INET6_ADDRSTRLEN is expected to be larger
-	 * than INET_ADDRSTRLEN, but this may be required in case
-	 * if for some unexpected reason IPv6 is not supported, and
-	 * INET6_ADDRSTRLEN is defined as 0
-	 * but this is not very likely and I am aware of no cases of
-	 * this in practice (editor)
-	 */
-	static char	addrstr[(INET6_ADDRSTRLEN > INET_ADDRSTRLEN ? INET6_ADDRSTRLEN : INET_ADDRSTRLEN) + 1];
-
-	if (!ai || !ai->ai_addr) {
-		errno = EINVAL;
-		return NULL;
-	}
-
-	addrstr[0] = '\0';
-	switch (ai->ai_family) {
-		case AF_INET: {
-			struct sockaddr_in	addr_in;
-			memcpy(&addr_in, ai->ai_addr, sizeof(addr_in));
-			return inet_ntop(AF_INET, &(addr_in.sin_addr), addrstr, INET_ADDRSTRLEN);
-		}
-		case AF_INET6: {
-			struct sockaddr_in6	addr_in6;
-			memcpy(&addr_in6, ai->ai_addr, sizeof(addr_in6));
-			return inet_ntop(AF_INET6, &(addr_in6.sin6_addr), addrstr, INET6_ADDRSTRLEN);
-		}
-		default:
-			errno = EAFNOSUPPORT;
-			return NULL;
-	}
-}
-
 /* return a pointer to the named ups if possible */
 upstype_t *get_ups_ptr(const char *name)
 {
@@ -853,7 +792,7 @@ static void client_connect(stype_t *server)
 
 	time(&client->last_heard);
 
-	client->addr = xstrdup(inet_ntopW(&csock));
+	client->addr = xstrdup(inet_ntopSS(&csock));
 
 	client->tracking = 0;
